fix(types): fixed types for model.init and sequelize.define; update docs
