+++ types/test/define.ts
@@ -1,4 +1,4 @@
-import { BuildOptions, DataTypes, Model } from 'sequelize';
+import { BuildOptions, DataTypes, Model, Optional } from 'sequelize';
 import { sequelize } from './connection';
 
 // I really wouldn't recommend this, but if you want you can still use define() and interfaces
@@ -8,26 +8,34 @@ interface UserAttributes {
   username: string;
   firstName: string;
   lastName: string;
-  createdAt: Date;
-  updatedAt: Date;
 }
 
-interface UserCreationAttributes extends Partial<UserAttributes> {}
+interface UserCreationAttributes extends Optional<UserAttributes, 'id'> {}
 
-interface UserModel extends Model<UserAttributes, UserCreationAttributes>, UserAttributes {}
+interface UserModel
+  extends Model<UserAttributes, UserCreationAttributes>,
+    UserAttributes {}
 
 const User = sequelize.define<UserModel>(
-  'User', { firstName: DataTypes.STRING }, { tableName: 'users' });
+  'User',
+  {
+    id: { type: DataTypes.NUMBER, primaryKey: true },
+    username: DataTypes.STRING,
+    firstName: DataTypes.STRING,
+    lastName: DataTypes.STRING,
+  },
+  { tableName: 'users' },
+);
 
 async function test() {
-    const user: UserModel = new User() as UserModel;
+  const user: UserModel = new User() as UserModel;
 
-    const user2: UserModel | null = await User.findOne();
-    if (!user2) return;
+  const user2: UserModel | null = await User.findOne();
+  if (!user2) return;
 
-    user2.firstName = 'John';
+  user2.firstName = 'John';
 
-    await user2.save();
+  await user2.save();
 }
 
 // The below doesn't define Attribute types, but should still work
@@ -36,9 +44,17 @@ interface UntypedUserModel extends Model, UserAttributes {}
 type UntypedUserModelStatic = typeof Model & {
   new (values?: keyof any, options?: BuildOptions): UntypedUserModel;
   customStaticMethod(): unknown;
-}
+};
 const UntypedUser = sequelize.define<UntypedUserModel>(
-  'User', { firstName: DataTypes.STRING }, { tableName: 'users' }) as UntypedUserModelStatic;
+  'User',
+  {
+    id: { type: DataTypes.NUMBER, primaryKey: true },
+    username: DataTypes.STRING,
+    firstName: DataTypes.STRING,
+    lastName: DataTypes.STRING,
+  },
+  { tableName: 'users' },
+) as UntypedUserModelStatic;
 
 UntypedUser.customStaticMethod = () => {};
