+++ types/test/typescriptDocs/ModelInit.ts
@@ -1,27 +1,6 @@
-# TypeScript
-
-Since v5, Sequelize provides its own TypeScript definitions. Please note that only TS >= 3.1 is supported.
-
-As Sequelize heavily relies on runtime property assignments, TypeScript won't be very useful out of the box. A decent amount of manual type declarations are needed to make models workable.
-
-## Installation
-
-In order to avoid installation bloat for non TS users, you must install the following typing packages manually:
-
-- `@types/node` (this is universally required in node projects)
-- `@types/validator`
-
-## Usage
-
-Example of a minimal TypeScript project with strict type-checking for attributes.
-
-<!--
-NOTE! 
-Keep the following code in sync with `typescriptDocs/ModelInit.ts` to ensure
-it typechecks correctly.
--->
-
-```ts
+/**
+ * Keep this file in sync with the code in the "Usage" section in typescript.md
+ */
 import {
   Sequelize,
   Model,
@@ -198,144 +177,3 @@ async function doStuffWithUser() {
   // the model or not
   console.log(ourUser.projects![0].name);
 }
-```
-
-### Usage without strict types for attributes
-
-The typings for Sequelize v5 allowed you to define models without specifying types for the attributes. This is still possible for backwards compatibility and for cases where you feel strict typing for attributes isn't worth it.
-
-<!--
-NOTE! 
-Keep the following code in sync with `typescriptDocs/ModelInitNoAttributes.ts` to ensure
-it typechecks correctly.
--->
-
-```ts
-import { Sequelize, Model, DataTypes } from 'sequelize';
-
-const sequelize = new Sequelize('mysql://root:asd123@localhost:3306/mydb');
-
-class User extends Model {
-  public id!: number; // Note that the `null assertion` `!` is required in strict mode.
-  public name!: string;
-  public preferredName!: string | null; // for nullable fields
-}
-
-User.init(
-  {
-    id: {
-      type: DataTypes.INTEGER.UNSIGNED,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    name: {
-      type: new DataTypes.STRING(128),
-      allowNull: false,
-    },
-    preferredName: {
-      type: new DataTypes.STRING(128),
-      allowNull: true,
-    },
-  },
-  {
-    tableName: 'users',
-    sequelize, // passing the `sequelize` instance is required
-  },
-);
-
-async function doStuffWithUserModel() {
-  const newUser = await User.create({
-    name: 'Johnny',
-    preferredName: 'John',
-  });
-  console.log(newUser.id, newUser.name, newUser.preferredName);
-
-  const foundUser = await User.findOne({ where: { name: 'Johnny' } });
-  if (foundUser === null) return;
-  console.log(foundUser.name);
-}
-```
-
-## Usage of `sequelize.define`
-
-In Sequelize versions before v5, the default way of defining a model involved using `sequelize.define`. It's still possible to define models with that, and you can also add typings to these models using interfaces.
-
-<!--
-NOTE! 
-Keep the following code in sync with `typescriptDocs/Define.ts` to ensure
-it typechecks correctly.
--->
-
-```ts
-import { Sequelize, Model, DataTypes, Optional } from 'sequelize';
-
-const sequelize = new Sequelize('mysql://root:asd123@localhost:3306/mydb');
-
-// We recommend you declare an interface for the attributes, for stricter typechecking
-interface UserAttributes {
-  id: number;
-  name: string;
-}
-
-// Some fields are optional when calling UserModel.create() or UserModel.build()
-interface UserCreationAttributes extends Optional<UserAttributes, 'id'> {}
-
-// We need to declare an interface for our model that is basically what our class would be
-interface UserInstance
-  extends Model<UserAttributes, UserCreationAttributes>,
-    UserAttributes {}
-
-const UserModel = sequelize.define<UserInstance>('User', {
-  id: {
-    primaryKey: true,
-    type: DataTypes.INTEGER.UNSIGNED,
-  },
-  name: {
-    type: DataTypes.STRING,
-  }
-});
-
-async function doStuff() {
-  const instance = await UserModel.findByPk(1, {
-    rejectOnEmpty: true,
-  });
-  console.log(instance.id);
-}
-```
-
-If you're comfortable with somewhat less strict typing for the attributes on a model, you can save some code by defining the Instance to just extend `Model` without any attributes in the generic types.
-
-<!--
-NOTE! 
-Keep the following code in sync with `typescriptDocs/DefineNoAttributes.ts` to ensure
-it typechecks correctly.
--->
-
-```ts
-import { Sequelize, Model, DataTypes } from 'sequelize';
-
-const sequelize = new Sequelize('mysql://root:asd123@localhost:3306/mydb');
-
-// We need to declare an interface for our model that is basically what our class would be
-interface UserInstance extends Model {
-  id: number;
-  name: string;
-}
-
-const UserModel = sequelize.define<UserInstance>('User', {
-  id: {
-    primaryKey: true,
-    type: DataTypes.INTEGER.UNSIGNED,
-  },
-  name: {
-    type: DataTypes.STRING,
-  },
-});
-
-async function doStuff() {
-  const instance = await UserModel.findByPk(1, {
-    rejectOnEmpty: true,
-  });
-  console.log(instance.id);
-}
-```
