+++ types/test/typescriptDocs/ModelInitNoAttributes.ts
@@ -1,0 +1,47 @@
+/**
+ * Keep this file in sync with the code in the "Usage without strict types for
+ * attributes" section in typescript.md
+ */
+import { Sequelize, Model, DataTypes } from 'sequelize';
+
+const sequelize = new Sequelize('mysql://root:asd123@localhost:3306/mydb');
+
+class User extends Model {
+  public id!: number; // Note that the `null assertion` `!` is required in strict mode.
+  public name!: string;
+  public preferredName!: string | null; // for nullable fields
+}
+
+User.init(
+  {
+    id: {
+      type: DataTypes.INTEGER.UNSIGNED,
+      autoIncrement: true,
+      primaryKey: true,
+    },
+    name: {
+      type: new DataTypes.STRING(128),
+      allowNull: false,
+    },
+    preferredName: {
+      type: new DataTypes.STRING(128),
+      allowNull: true,
+    },
+  },
+  {
+    tableName: 'users',
+    sequelize, // passing the `sequelize` instance is required
+  },
+);
+
+async function doStuffWithUserModel() {
+  const newUser = await User.create({
+    name: 'Johnny',
+    preferredName: 'John',
+  });
+  console.log(newUser.id, newUser.name, newUser.preferredName);
+
+  const foundUser = await User.findOne({ where: { name: 'Johnny' } });
+  if (foundUser === null) return;
+  console.log(foundUser.name);
+}
